[Plugins] Move client components to separate files
